requirements.txt
discord.py[voice]
wavelink
PyNaCl
pytz
python-dotenv
aiosqlite